Refactoring in TrendingRecipeCell
