+++ BestRecipes/Extensions/UIFont+.swift
@@ -1,0 +1,13 @@
+//
+//  UIFont+.swift
+//  BestRecipes
+//
+//  Created by Илья Шаповалов on 08.09.2023.
+//
+
+import UIKit
+
+extension UIFont {
+    static let titleFont: UIFont? = .init(name: "Arial-BoldMT", size: 16)
+    static let creatorTitleFont: UIFont? = .init(name: "ArialMT", size: 12)
+}
